Fix compiler warning
git-svn-id: https://svn.smr.arpa.emr.it/arkimet/arkimet/trunk@1553 9b5092fc-ae44-42f6-b796-0d08dad2e17b

+++ arki/metadata.cc
@@ -286,7 +286,7 @@ void Metadata::write(int outfd, const std::string& filename) const
 	ssize_t res = ::write(outfd, encoded.data(), encoded.size());
 	if (res < 0)
 		throw wibble::exception::System("writing metadata to " + filename);
-	if (res != encoded.size())
+	if ((size_t)res != encoded.size())
 		throw wibble::exception::Consistency("writing metadata to " + filename, "written only " + str::fmt(res) + " bytes out of " + str::fmt(encoded.size()));
 
 	// If the source is inline, then the data follows the metadata
@@ -295,7 +295,7 @@ void Metadata::write(int outfd, const std::string& filename) const
 		ssize_t res = ::write(outfd, m_inline_buf.data(), m_inline_buf.size());
 		if (res < 0)
 			throw wibble::exception::System("writing data to " + filename);
-		if (res != m_inline_buf.size())
+		if ((size_t)res != m_inline_buf.size())
 			throw wibble::exception::Consistency("writing data to " + filename, "written only " + str::fmt(res) + " bytes out of " + str::fmt(m_inline_buf.size()));
 	}
 }
